LPS-65312 Move
