+++ modules/apps/forms-and-workflow/portal-workflow/portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/upgrade/KaleoServiceUpgrade.java
@@ -17,9 +17,9 @@ package com.liferay.portal.workflow.kaleo.upgrade;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 import com.liferay.portal.workflow.kaleo.upgrade.v1_0_0.UpgradeKaleoTaskInstanceToken;
 import com.liferay.portal.workflow.kaleo.upgrade.v1_1_0.UpgradeWorkflowContext;
-import com.liferay.portal.workflow.kaleo.upgrade.v1_2_0.UpgradeKaleoLog;
-import com.liferay.portal.workflow.kaleo.upgrade.v1_2_0.UpgradeKaleoNotificationRecipient;
 import com.liferay.portal.workflow.kaleo.upgrade.v1_2_0.UpgradeSchema;
+import com.liferay.portal.workflow.kaleo.upgrade.v1_2_1.UpgradeKaleoLog;
+import com.liferay.portal.workflow.kaleo.upgrade.v1_2_1.UpgradeKaleoNotificationRecipient;
 import com.liferay.portal.workflow.kaleo.upgrade.v1_3_0.UpgradeClassNames;
 import com.liferay.portal.workflow.kaleo.upgrade.v1_3_0.UpgradeKaleoAction;
 import com.liferay.portal.workflow.kaleo.upgrade.v1_3_0.UpgradeKaleoDefinition;
@@ -44,11 +44,14 @@ public class KaleoServiceUpgrade implements UpgradeStepRegistrator {
 
 		registry.register(
 			"com.liferay.portal.workflow.kaleo.service", "1.1.0", "1.2.0",
-			new UpgradeSchema(), new UpgradeKaleoLog(),
-			new UpgradeKaleoNotificationRecipient());
+			new UpgradeSchema());
 
 		registry.register(
-			"com.liferay.portal.workflow.kaleo.service", "1.2.0", "1.3.0",
+			"com.liferay.portal.workflow.kaleo.service", "1.2.0", "1.2.1",
+			new UpgradeKaleoLog(), new UpgradeKaleoNotificationRecipient());
+
+		registry.register(
+			"com.liferay.portal.workflow.kaleo.service", "1.2.1", "1.3.0",
 			new UpgradeClassNames(), new UpgradeKaleoAction(),
 			new UpgradeKaleoDefinition());
 	}
